Makes the Server side handle demand
This makes the server side handle demand by setting the actual demand on the Client side to 0, it also Sync the displayed demand from the server to the client.

When syncing of Zones are introduced this will make Demand and RCI work like in the normal game.

Also, i Opted for only syncing the displayed UI demand when the Residential Demand is updated, it should not make a noticable difference but should be less resouce demanding since its only sending one command with each demand uppdate instead of three.