This Sync the Node and SegmentNodes (Roads, Waterpipes, powerlines ect)
This adds a new extension that syncs Nodes and SegmentNodes, it uses the ThreadingExtensionBase and use a new Thread to check for changes in the Nodes and sending these Changes.

For the Extension to function, the reciever needs to first recieve all the Nodes before recieving the NodesSegment. I have tried to ensure this by adding Thread.Sleep and using locks, but this is in no way a "safe" implementation that ensures that all is recieved and handled in the right order. Before release we need a new and more stable way of handling this.