Remove needs for dictionary, and greatly improve the sync of Nodes and Segments
This removes the need for dictionaries (it's still used to handle zones)
This update should greatly improve the sync of segments and nodes.
but there are still some problem with how the waterpipes are syncing since they are a combination of buildings and Node segments.

You can't connect a waterpipe to a watertower, but you can place a watertower on the waterpipe.

For playing i would recommend building the watertower and sewage and connect both things to a pipe saving and then sharing the save.