Sync Creation, modification and release of districts
This syncs the Creation, modification and release of districts. It does not sync policies.

Also it contains a small bug the name of the district will not show on creation , to make the name update the reciver of the district has to click on it using the district brush, the same is true for updating the names position and removing the name.